- Add Transaction class. - Add TestUtils class with test method to create an arbitrary number of transactions. - Add 1st version of rowlayout for the ListView. - Modified main layout to be a Listview. - Added adapter for transaction that uses rowlayout. - Changed amount type from float to double. - Tweak xmls to make content look acceptable (as opposed to make it look good).

+++ app/src/main/java/com/personal/rprs/cashtracker/Transaction.java
@@ -7,14 +7,14 @@ import java.util.Date;
  */
 public class Transaction {
 
-    public float amount;
+    public double amount;
     public String description;
     public Date date;
     public boolean cc;
     public boolean debtWithRoomie;
 
     public Transaction(
-            float inAmount,
+            double inAmount,
             String inDescription,
             Date inDate,
             boolean inCc,
